pytest_tipsi plugin
