--- tipsi_tools/testing/pytest/README.md
@@ -1,11 +1,0 @@
-This module contains useful pytest fixtures for drf and django
-
-## Usage
-Add to `conftest.py`
-```
-from tipsi_tools.testing.pytest import *
-```
-
-
-## TODO
-Add tests
